requirements.txt
fastapi
motor
uvicorn
streamlit
requests
dotenv
google-generativeai
